Fix compiler warnings

+++ src/command_charge.cpp
@@ -218,7 +218,7 @@ void Command::exportRecurringCharges() {
 
     vector<JRecord> records;
 
-    for (int i = 0;i < results.size();i++) {
+    for (size_t i = 0;i < results.size();i++) {
         DBAccount account = results.at(i);
 
         JRecord r = account.getRecord();
