+++ src/db_account.h
@@ -106,7 +106,7 @@ class DBAccount : public DBEntity {
 
             os << getDeleteAllStatement() << endl;
 
-            for (int i = 0;i < results.size();i++) {
+            for (size_t i = 0;i < results.size();i++) {
                 os << results[i].getInsertStatement() << endl;
             }
